Revamp real Schur code, avoid underflow
This change is mainly a substantial revision of the basic real Schur
code. The old one had troubles with some edge cases, particularly when
underflow occurred.  The replacement is intended to be a faithful
translation from LAPACK, and uses Householder reflectors also based
on LAPACK.

The non-standard real Schur form is no longer included.

Unused code blocks were removed, and the test suite was enhanced.